update zip_code data type

+++ retail_analytics/snapshots/snapshot_customers.sql
@@ -21,7 +21,7 @@
             customer_id,
             customer_address AS address,
             customer_state AS state,
-            customer_zip_code::integer AS zip_code_prefix,
+            customer_zip_code AS zip_code_prefix,
             customer_created_date::timestamp AS created_date,
             COALESCE(customer_updated_date, customer_created_date::timestamp) AS updated_date,
             ROW_NUMBER() OVER (PARTITION BY customer_id) AS row_num
